+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Improved shard health check to be more accurate.
+
 ## [v17.7.3](https://github.com/lexicalunit/spellbot/releases/tag/v17.7.3) - 2026-01-15
 
 ### Changed